Fixed a bug where deleting a card from one user would result in deleting the same card from other users. Fixed a bug where when setting a level for a custom card, the corresponding counter was not installed. Added the ability to delete a user card.

+++ PHP/deleteCard.php
@@ -12,7 +12,7 @@
         }
         $userId = $_SESSION['userId'];
         $cardId = $_POST['cardId'];
-        $sql = "DELETE FROM UserDictionary WHERE IdDictionary = $cardId"; // SQL запрос
+        $sql = "DELETE FROM UserDictionary WHERE IdDictionary = $cardId AND IdUser = $userId"; // SQL запрос
         $result = mysqli_query($Connect, $sql); // выполнение запроса
         mysqli_close($Connect); // Закрываем соединение с базой данных
     }
